feat: use loop time rather than wall-clock time
We now use the time base of the async loop rather than time.time() clock
time. The loop uses time.monotonic, which has an arbitrary epoch. We
need to fast-forward the loop's time to test sleep/timeout without the
tests waiting for real time to pass. This is easier if the code is using
loop time as its reference.