feat!: make open_kv() async
It needs to create an aiohttp.ClientSession() and while that can be done
synchronously, doing so requires an event loop to be running. Making it
async should avoid a usage error where people assume it's independent of
a loop, and also allows it to work naturally in Python REPLs like
iPython, where sync methods don't run in the context of an event loop,
but awaited calls do. It's still possible to manually construct Kv()
synchronously if really required.