test: use strict BigInt/Number V8 decoder in MockKvDb
We now use a customised V8 decoder for the MockKvDb API that mirrors the
default encoder used by Kv (that encodes int as BigInt): it always
decodes int32 values as float rather than the v8serialize default of
decoding int32 as int and Double as float.

We don't rely on this being in place as we currently encode all int32
range values as Double (which are always decoded as float), but it seems
prudent to ensure that we don't mistakenly treat small ints as BigInt
instead of Number.