feat!: make Result.flatten() idempotent
Result.flatten() now returns the Ok as is if there's no nested Result.

It was raising TypeError if Ok did not hold a Result. In contrast, the
Option.flatten() method is idempotent, returning the Some as is if
there's no nested Option. I think this approach is preferable to
raising.