refactor: make Result a base class rather than Union
Same approach as the Option base. Compared to a Union, this has the
benefit of providing a namespace to hold static functions, and a
constructor. Also, mypy can fail to infer the type of literals
containing Ok and Err values, resulting in object, or the ResultMethods
prototype. By using a common base class, it should infer the base type
when it would otherwise fall back to object.