refactor!: make Option and Result unions again
When they share a parent in the type system there seem to be edge cases
where the types are not narrowed/excluded correctly. Using a union of
two disjoint types seems to give type checkers an easier time.

One slight downside is that in some cases literal expressions (like
lists) containing both sides of the union get inferred as object. These
need to be given explicit annotations. But in practice, normally the
result of a function provides such an annotation automatically.