perf: avoid redundant encoding in pack_key_range()
When the same key tuple is used as the start and end of a range, we now
only pack the tuple once. This happens either if the same key tuple is
the start and end, or a prefix is used without a start/end.

I did some quick benching of the effect of using lru_cache on pack() and
the benefit is very marginal in the ideal case where the same key is
repeatedly packed. I don't think it's worth it.