feat: improve KvKey implementation
KvKey now has equality and order comparison behaviour that follows the
packed key representation. (Plain Python tuples compare differently
to packed keys, because int and float compare/hash the same as Python
values, but not when packed.)

KvKey now caches the packed representation the first time it's
calculated, so each instance only packs once. We use the packed
representation in various other methods now.

The tuple methods and type annotations are also more accurate, as they
are overridden to return KvKey instances, not plain tuples. It's no
longer an actual tuple subclass, rather it's a regular class that is
annotated to pretend to be a tuple subclass in order to use some of the
special tuple type handling.