perf: memoise recent pack_key() calls
The pack_key() function now caches the most recent 128 keys. I
experimented previously with functools.lru_cache(), but it adds about as
much overhead as just encoding the tuple, and also doesn't handle the
difference between float and int values in the key argument tuple, so
that is no good.

We use a purpose-specific cache here that calculates a cache key value
for the key tuple that includes types for each position to account for
the int/float distinction. The overhead of this is about 1:1 when
packing a 1-2 element tuple and is consistently fast with larger key
tuples, while actually packing grows in clock time linearly with length.