refactor!: remove customised v8serialize Encoder
denokv no longer customises the v8serialize Encoder/Decoder. Previously
it changed the encoding behaviour to always treat int as bigint. This
was necessary because v8serialize did not have a specific type for
bigint, so int would get encoded as float (JS Number) or bigint
depending on the size of the int, so there was no way to encode a small
int as a bigint.

v8serialize now includes a JSBigInt type which solves this problem.