feat: add types for Kv write API
PlannedWrite is the central type, it represents the collection of
mutation operations to be applied to the DB in a single atomic write.

We have a KvWriter protocol which PlannedWrite implements, it'll be used
by Kv so that we don't have a circular dependency between denokv.kv.Kv
and denokv._kv_writes.PlannedWrite.

In addition, we now have a ProtobufMessageRepresentation[MessageT] type
which can be used to generalise the types representing datapath protocol
types, like Sum, Max, Min, etc; to avoid coupling PlannedWrite to our
default implementations.