feat: rework PlannedWrite & Mutation APIs
The PlannedWrite and Mutation types now have much better support and
APIs for performing atomic number operations, like sum, min and max.

The first main feature is that the 3 number types (bigint, u64, float) can
be selected explicitly with a number_type parameter, or implicitly via
the type of the number value, and both options are type-safe. The second
is that we support min/max mutations on bigint and float types, not just
u64 (KvU64), and we support sum on KvU64 with both wrapping and clamp
limits. The underlying datapath protocol (and Deno's own KV API) do not
support these. This makes the atomic number APIs more intuitive, as
otherwise the user needs to know about the arbitrary restriction that
only certain number types can use Min/Max/Sum.

I don't love the massive amount of @overload annotations it takes to
type-annotate these functions well, but I don't think there's a better
alternative.

The way that PlannedWrite interfaces with the main Kv class is now
decoupled with a general interface, so Kv and PlannedWrite do not have
explicit dependencies on each other. This makes it easier to test and
modularise the codebase, and also allows alternative writer
implementations to be created for special cases.